+++ src/index.tsx
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 import Worker from 'worker-loader!./workers/web'
 import App from './App'
-import { I18nProvider } from './modules/i18n'
+import { I18nProvider } from './contexts/i18n'
 import createStore from './store'
 import './styles/global.css'
 import debug from './utils/debug'
